Forsett å ta i bruk nav-faker
Også gjort litt refaktorering for familierelasjoner. Flyttet kode ut i
egne filer.
