+++ src/mock/utils/fnr-utils.ts
@@ -1,5 +1,3 @@
-import *  as moment from 'moment';
-
 import navfaker from 'nav-faker/dist/index';
 import { Kjønn as Kjønnstyper } from 'nav-faker/dist/fodselsnummer/fodselsnummer';
 import { Kjønn } from '../../models/person/person';
@@ -8,19 +6,16 @@ export function randomFodselsnummer(): string {
     return navfaker.fødselsnummer.generer();
 }
 
-export function seededTilfeldigFodselsnummer(minAlder: number, maxAlder: number, kjønn?: Kjønn) {
-    const fromDate = moment().subtract(maxAlder, 'years').toDate();
-    const toDate = moment().subtract(minAlder, 'years').toDate();
-    const tilfeldigFødselsdato = navfaker.dato.mellom(fromDate, toDate);
-    return navfaker.fødselsnummer.generer({fødselsdato: tilfeldigFødselsdato, kjønn: getKjønnstype(kjønn)});
+export function tilfeldigFodselsnummer(date: Date, kjønn?: Kjønn) {
+    return navfaker.fødselsnummer.generer({fødselsdato: date, kjønn: getKjønnstype(kjønn)});
 }
 
 function getKjønnstype(kjønn?: Kjønn): Kjønnstyper | undefined {
     if (kjønn) {
         if (kjønn === Kjønn.Kvinne) {
-            return Kjønnstyper.KVINNE;
+            return 0;
         } else {
-            return Kjønnstyper.MANN;
+            return 1;
         }
     } else {
         return undefined;
